onyx: camera: Add OnePlus modification

Conflicts:
	QCamera2/stack/mm-camera-test/src/mm_qcamera_dual_test.c
	camera/QCamera2/HAL/Android.mk.orig
	camera/QCamera2/HAL/QCamera2Factory.cpp
	camera/QCamera2/HAL/QCamera2HWI.cpp
	camera/QCamera2/HAL/QCamera2HWI.h
	camera/QCamera2/HAL/QCamera2HWICallbacks.cpp
	camera/QCamera2/HAL/QCameraMem.cpp
	camera/QCamera2/HAL/QCameraMem.h
	camera/QCamera2/HAL/QCameraParameters.cpp
	camera/QCamera2/HAL/QCameraParameters.h
	camera/QCamera2/HAL/QCameraStream.cpp
	camera/QCamera2/stack/common/cam_types.h
	camera/QCamera2/stack/common/mm_camera_interface.h
	camera/QCamera2/stack/common/mm_jpeg_interface.h
	camera/QCamera2/stack/mm-camera-interface/inc/mm_camera.h
	camera/QCamera2/stack/mm-camera-interface/src/mm_camera.c
	camera/QCamera2/stack/mm-camera-interface/src/mm_camera_channel.c
	camera/QCamera2/stack/mm-camera-interface/src/mm_camera_stream.c
	camera/QCamera2/stack/mm-jpeg-interface/src/mm_jpeg.c
	camera/QCamera2/stack/mm-jpeg-interface/src/mm_jpeg_exif.c
	camera/QCamera2/stack/mm-jpeg-interface/src/mm_jpeg_interface.c
	camera/QCamera2/util/QCameraCmdThread.cpp

+++ camera/QCamera2/HAL/QCameraMem.cpp
@@ -1250,7 +1250,7 @@ int QCameraVideoMemory::allocate(int count, int size)
         }
         nh->data[0] = mMemInfo[i].fd;
         nh->data[1] = 0;
-        nh->data[2] = (int)mMemInfo[i].size;
+        nh->data[2] = mMemInfo[i].size;
         nh->data[3] = 0;
         nh->data[4] = 0;
     }
